refactor(githooks-manager): unified addHook API on GitHooksManager base class
- Add static of() and abstract addHook() to GitHooksManager so callers
  need only one import instead of branching on Husky vs Lefthook
- Implement addHook() in Husky (delegates to createHook) and Lefthook
  (delegates to addCommand)
- Extract GitClientHook/GitHooksManagerType enums into types.ts and
  re-export from barrel
- Move lint-staged into husky/lint-staged/ sub-directory
- Fix addScript() to handle an existing action that has no scripts yet
- Update jest.ts and vitest.ts to use GitHooksManager.of().addHook()
- Use runTest() in vitest.ts to cover all NodePackageManager values
- Add jest: false to lefthook snapshot test that tests a Lefthook-only config
