+++ src/components/githooks-manager/husky/index.ts
@@ -1,4 +1,3 @@
 export * from "./component";
 export * from "./hook-file";
-export * from "./lintstaged";
-export * from "./lintstagedrule";
+export * from "./lint-staged";
