+++ src/components/githooks-manager/husky/lint-staged/rule.ts
