+++ src/components/githooks-manager/index.ts
@@ -2,3 +2,4 @@ export * from "./action";
 export * from "./husky";
 export * from "./lefthook";
 export * from "./manager";
+export * from "./types";
