+++ src/components/githooks-manager/lefthook/component.ts
@@ -1,7 +1,8 @@
 import type { Component, Project } from "projen";
 import { NodePackageManager } from "projen/lib/javascript";
 import type { GitHooksEnabledProject } from "../../../typescript/githooks-enabled-project";
-import { GitClientHook, GitHooksManager } from "../manager";
+import { GitHooksManager } from "../manager";
+import { GitClientHook } from "../types";
 import { LefthookCommand, type LefthookCommandOptions } from "./command";
 import { LefthookConfig } from "./config";
 import { LefthookFile } from "./file";
@@ -46,6 +47,10 @@ export class Lefthook extends GitHooksManager {
 		});
 	}
 
+	public addHook(hook: GitClientHook, command: string): void {
+		this.addCommand(hook, { run: command });
+	}
+
 	public addCommand(hookName: GitClientHook, command: LefthookCommandOptions) {
 		const action = this.config.actions?.find(
 			(hook) => hook.actionName === hookName,
@@ -62,18 +67,24 @@ export class Lefthook extends GitHooksManager {
 	}
 
 	public addScript(hookName: GitClientHook, script: LefthookScriptOptions) {
-		const action = this.config.actions?.find(
+		const index = this.config.actions.findIndex(
 			(hook) => hook.actionName === hookName,
 		);
 
-		if (!action) {
+		if (index === -1) {
 			this.config.actions.push({
 				actionName: hookName,
 				scripts: [script],
 			});
+			return;
 		}
 
-		action?.scripts?.push(script);
+		const action = this.config.actions[index];
+		if (action.scripts) {
+			action.scripts.push(script);
+		} else {
+			this.config.actions[index] = { ...action, scripts: [script] };
+		}
 	}
 
 	preSynthesize(): void {
